# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Studio Kanban API",
    "description": "REST API for the Studio Kanban music production board",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
...
      "Song": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "projectId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "Track 1"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "Upbeat intro track"
          },
          "order": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01 12:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01 12:00:00"
          }
        },
        "required": [
          "id",
          "projectId",
          "title",
          "description",
          "order",
          "createdAt",
          "updatedAt"
        ]
      },
      "Subtask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440004"
          },
          "taskId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440003"
          },
          "title": {
            "type": "string",
            "example": "Set up microphones"
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "order": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01 12:00:00"
          }
        },
        "required": [
          "id",
          "taskId",
          "title",
          "completed",
          "order",
          "createdAt"
        ]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440003"
          },
          "songId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "columnId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "title": {
            "type": "string",
            "example": "Record drums"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "example": "Record kick and snare tracks"
          },
          "assignee": {
            "type": [
              "string",
              "null"
            ],
            "example": "Alice"
          },
          "order": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01 12:00:00"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01 12:00:00"
          },
          "subtasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subtask"
            },
            "description": "Subtasks for this task"
          }
        },
        "required": [
          "id",
          "songId",
          "columnId",
          "title",
          "description",
          "assignee",
          "order",
          "createdAt",
          "updatedAt",
          "subtasks"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/songs/{id}": {
      "put": {
        "tags": [
          "Songs"
        ],
        "summary": "Update a song",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "550e8400-e29b-41d4-a716-446655440002"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Updated Title"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Songs"
        ],
        "summary": "Delete a song",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "550e8400-e29b-41d4-a716-446655440002"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}/order": {
      "patch": {
        "tags": [
          "Songs"
        ],
        "summary": "Update song order",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "550e8400-e29b-41d4-a716-446655440002"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "integer",
                    "example": 2
                  }
                },
                "required": [
                  "order"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/songs/{songId}/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List tasks for a song (with subtasks)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "550e8400-e29b-41d4-a716-446655440002"
            },
            "required": true,
            "name": "songId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks with subtasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create a task for a song",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "550e8400-e29b-41d4-a716-446655440002"
            },
            "required": true,
            "name": "songId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "columnId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Record drums"
                  },
                  "description": {
                    "type": "string",
                    "example": "Record kick and snare"
                  },
                  "assignee": {
                    "type": "string",
                    "example": "Alice"
                  }
                },
                "required": [
                  "columnId",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}